requirements.txt
pylint
numpy
f90nml==1.4.4
torch
flake8 
matplotlib
pytest 
PyYAML
dask[complete]
dask-jobqueue
coverage
# python==3.11
scikit-learn==1.3.1
